Fix cast warning

+++ ouzel/graphics/raspbian/RendererOGLRasp.cpp
@@ -99,7 +99,7 @@ namespace ouzel
                 EGL_GREEN_SIZE, 8,
                 EGL_BLUE_SIZE, 8,
                 EGL_ALPHA_SIZE, 8,
-                EGL_DEPTH_SIZE, newDepthBits,
+                EGL_DEPTH_SIZE, static_cast<int>(newDepthBits),
                 EGL_SURFACE_TYPE, EGL_WINDOW_BIT,
                 EGL_NONE
             };
